Rename Brassica::Api to Api and move it to app/services
