+++ spec/models/model_spec.rb
@@ -86,7 +86,7 @@ RSpec.describe ActiveRecord::Base do
   end
 
   context "supporting API" do
-    Brassica::Api.readable_models.each do |klass|
+    Api.readable_models.each do |klass|
       context klass do
         it "includes Filterable" do
           expect(klass.ancestors).to include(Filterable)
